+++ Wire-iOS/Sources/UserInterface/Components/CGFloat+Const.swift
@@ -34,6 +34,10 @@ extension StyleKitIcon.Size {
 }
 
 extension CGFloat {
+    enum WipeCompletion {
+        static let buttonHeight: CGFloat = 48
+    }
+
     enum PasscodeUnlock {
         static let textFieldHeight: CGFloat = 40
         static let buttonHeight: CGFloat = 40
